+++ docs/src/App.tsx
@@ -60,6 +60,7 @@ const SnackbarDemoPage = page(() => import('./pages/SnackbarDemo'));
 const TreeDemoPage = page(() => import('./pages/TreeDemo'));
 const DropzoneDemoPage = page(() => import('./pages/DropzoneDemo'));
 const DateSelectorDemoPage = page(() => import('./pages/DateSelectorDemo'));
+const CodeBlockDemoPage = page(() => import('./pages/CodeBlockDemo'));
 const MarkdownDemoPage = page(() => import('./pages/MarkdownDemo'));
 const OverviewPage = page(() => import('./pages/Overview'));
 const QuickstartPage = page(() => import('./pages/Quickstart'));
@@ -315,6 +316,10 @@ export function App() {
         <Route
           path='/dateselector-demo'
           element={<DateSelectorDemoPage />}
+        />
+        <Route
+          path='/codeblock-demo'
+          element={<CodeBlockDemoPage />}
         />
         <Route
           path='/markdown-demo'